Fixed all warnings

+++ game/game_libs/cl_dlls/ammohistory.cpp
@@ -57,7 +57,7 @@ void HistoryResource::AddToHistory(int iType, int iId, int iCount)
 	}
 
 	HIST_ITEM* freeslot = &rgAmmoHistory[iCurrentHistorySlot++];  // default to just writing to the first slot
-	HISTORY_DRAW_TIME = CL_CvarGetFloat("hud_drawhistory_time");
+	HISTORY_DRAW_TIME = static_cast<int>(CL_CvarGetFloat("hud_drawhistory_time"));
 
 	freeslot->type = iType;
 	freeslot->iId = iId;
@@ -89,7 +89,7 @@ void HistoryResource::AddToHistory(int iType, const char* szName, int iCount)
 	freeslot->type = iType;
 	freeslot->iCount = iCount;
 
-	HISTORY_DRAW_TIME = CL_CvarGetFloat("hud_drawhistory_time");
+	HISTORY_DRAW_TIME = static_cast<int>(CL_CvarGetFloat("hud_drawhistory_time"));
 	freeslot->DisplayTime = gHUD.m_flTime + HISTORY_DRAW_TIME;
 }
