+++ game/game_libs/game_shared/weapons/genericweapon.h
@@ -70,9 +70,9 @@ protected:
 	const T* GetAttackModeFromAttributes(uint32_t index) const
 	{
 		const WeaponAtts::WACollection& atts = WeaponAttributes();
-		ASSERT(index < atts.AttackModes.Count());
+		ASSERT(index < static_cast<uint32_t>(atts.AttackModes.Count()));
 
-		if ( index >= atts.AttackModes.Count() )
+		if ( index >= static_cast<uint32_t>(atts.AttackModes.Count()) )
 		{
 			return nullptr;
 		}
@@ -226,9 +226,9 @@ inline const WeaponAtts::WABaseAttack* CGenericWeapon::GetAttackModeFromAttribut
 	uint32_t index) const
 {
 	const WeaponAtts::WACollection& atts = WeaponAttributes();
-	ASSERT(index < atts.AttackModes.Count());
+	ASSERT(index < static_cast<uint32_t>(atts.AttackModes.Count()));
 
-	if ( index >= atts.AttackModes.Count() )
+	if ( index >= static_cast<uint32_t>(atts.AttackModes.Count()) )
 	{
 		return nullptr;
 	}
